[refactor]: 피드 댓글 화면에서 프로필 누르면 사용자 피드로 이동하게 수정 (#101)

+++ app/src/main/java/com/texthip/thip/data/model/comments/response/CommentsCreateResponse.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CommentsCreateResponse(
     val commentId: Int?,
-    val creatorId: Int?,
+    val creatorId: Long?,
     val creatorProfileImageUrl: String?,
     val creatorNickname: String?,
     val aliasName: String?,
